pdf2markdown.py
from marker.converters.pdf import PdfConverter
from marker.models import create_model_dict
from marker.output import text_from_rendered

converter = PdfConverter(
    artifact_dict=create_model_dict(),
)
rendered = converter("data/Hands-On_Large_Language_Models.pdf")
text, _, images = text_from_rendered(rendered)