pyproject.toml
[project]
name = "streamlit-rag-llm-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.10.1",
    "marker-pdf>=0.1.3",
    "streamlit>=1.50.0",
    "torch==2.2.2",
    "transformers>=4.57.1",
]
